Added Visual Studio 2017 project files, removed 2013 and below (#7)
* Added project files for Visual Studio 2017
* Updated readme and premake for vs2017
* Removed VS2010, 2012, 2013

+++ amd_geometryfx/src/AMD_GeometryFX_Filtering.cpp
@@ -200,7 +200,7 @@ public:
             lastTriangle = std::min (
                 firstTriangle + SmallBatchMergeConstants::BATCH_SIZE, request.mesh->faceCount);
 
-            assert (currentCluster < request.mesh->clusters.size ());
+            assert (currentCluster < static_cast<int>(request.mesh->clusters.size ()));
             const auto& clusterInfo = request.mesh->clusters[currentCluster];
             ++currentCluster;
